requirements.txt
streamlit
pillow
pytesseract
transformers
gtts
torch
torchvision
torchaudio

